// backend/src/types/Customer.ts
export interface Customer {
  id: number;
  name: string;
  phone: number;
  email: string;
  created_at: string;
  updated_at: string;
}
import pool from "../config/db";

export interface Customer {
  id: number;
  name: string;
  phone: number;
  email: string;
  created_at: string;
  updated_at: string;
}